+++ lib/src/destination.dart
@@ -60,14 +60,16 @@ class Destination<T extends DestinationParameters> {
   /// Creates a destination that provides a navigator with nested destinations.
   ///
   /// An optional [builder] parameter is basically the same like normal [Destination.builder],
-  /// but has additional [child] parameter, which contains the nested content that built by [navigator].
+  /// but has additional [childBuilder] parameter, which provides the nested content that built by [navigator].
   /// The implementation of [builder] function must include this child widget sub-tree
   /// in the result for correct displaying the nested content.
   ///
   Destination.transit({
     required this.path,
     required this.navigator,
-    Widget Function(BuildContext context, T? parameters, Widget child)? builder,
+    Widget Function(
+            BuildContext context, T? parameters, WidgetBuilder childBuilder)?
+        builder,
     this.isHome = false,
     this.redirections = const <Redirection>[],
     this.tag,
@@ -144,7 +146,8 @@ class Destination<T extends DestinationParameters> {
   final Future<Destination?> Function(Destination<T> destination)?
       upwardDestinationBuilder;
 
-  late final Widget Function(BuildContext context, T? parameters, Widget child)?
+  late final Widget Function(
+          BuildContext context, T? parameters, WidgetBuilder childBuilder)?
       _transitBuilder;
 
   /// Whether this destination is final, i.e. it builds a content
@@ -177,11 +180,11 @@ class Destination<T extends DestinationParameters> {
     if (isFinalDestination) {
       return builder!(context, parameters);
     } else {
-      final nestedContent = navigator!.build(context);
       if (_transitBuilder != null) {
-        return _transitBuilder!(context, parameters, nestedContent);
+        return _transitBuilder!(context, parameters,
+            (nestedContext) => navigator!.build(nestedContext));
       } else {
-        return nestedContent;
+        return navigator!.build(context);
       }
     }
   }
@@ -264,14 +267,14 @@ class Destination<T extends DestinationParameters> {
 /// [quiet] - replace the previous destination with the current one without animations.
 ///
 /// See also:
-/// - [DestinationAction]
+/// - [TransitionMethod]
 /// - [DestinationTransition]
 ///
 class DestinationSettings {
   /// Creates an instance of [DestinationSettings].
   ///
   const DestinationSettings({
-    required this.action,
+    required this.transitionMethod,
     required this.transition,
     this.redirectedFrom,
     this.reset = false,
@@ -300,9 +303,9 @@ class DestinationSettings {
   /// How the destination will update the navigation stack.
   ///
   /// See also:
-  ///  - [DestinationAction]
+  ///  - [TransitionMethod]
   ///
-  final DestinationAction action;
+  final TransitionMethod transitionMethod;
 
   /// Visual effects that would be applied on updating the stack with the destination.
   ///
@@ -341,13 +344,13 @@ class DestinationSettings {
   ///
   DestinationSettings copyWith({
     // TODO: Add other properties
-    DestinationAction? action,
+    TransitionMethod? transitionMethod,
     Destination? redirectedFrom,
     bool? reset,
     bool? updateHistory,
   }) =>
       DestinationSettings(
-        action: action ?? this.action,
+        transitionMethod: transitionMethod ?? this.transitionMethod,
         transition: transition,
         redirectedFrom: redirectedFrom ?? this.redirectedFrom,
         reset: reset ?? this.reset,
@@ -359,7 +362,7 @@ class DestinationSettings {
 class _DefaultDestinationSettings extends DestinationSettings {
   const _DefaultDestinationSettings()
       : super(
-          action: DestinationAction.push,
+          transitionMethod: TransitionMethod.push,
           transition: DestinationTransition.material,
         );
 }
@@ -367,7 +370,7 @@ class _DefaultDestinationSettings extends DestinationSettings {
 class _DialogDestinationSettings extends DestinationSettings {
   const _DialogDestinationSettings()
       : super(
-          action: DestinationAction.push,
+          transitionMethod: TransitionMethod.push,
           transition: DestinationTransition.materialDialog,
         );
 }
@@ -375,14 +378,14 @@ class _DialogDestinationSettings extends DestinationSettings {
 class _QuietDestinationSettings extends DestinationSettings {
   const _QuietDestinationSettings()
       : super(
-          action: DestinationAction.replace,
+          transitionMethod: TransitionMethod.replace,
           transition: DestinationTransition.none,
         );
 }
 
-/// An action that is used to update the navigation stack with the destination.
+/// A way of transition to the destination.
 ///
-enum DestinationAction {
+enum TransitionMethod {
   /// The destination will be added to the navigation stack.
   /// On navigation back, the destination will be removed from the stack
   /// and previous destination will be restored.