+++ test/common/common.dart
@@ -14,11 +14,22 @@ class TestDestinations {
     builder: dummyBuilder,
     isHome: true,
   );
+  static final homeRedirectionApplied = Destination(
+    path: '/home',
+    builder: dummyBuilder,
+    isHome: true,
+    redirections: [
+      Redirection(
+        destination: TestDestinations.login,
+        validator: (destination) async => false,
+      ),
+    ],
+  );
   static final about = Destination(
     path: '/settings/about',
     builder: dummyBuilder,
   );
-  static final aboutWithRedirection = Destination(
+  static final aboutRedirectionNotApplied = Destination(
     path: '/settings/about',
     builder: dummyBuilder,
     redirections: [
@@ -29,7 +40,7 @@ class TestDestinations {
       ),
     ],
   );
-  static final aboutWithInvalidRedirection = Destination(
+  static final aboutRedirectionApplied = Destination(
     path: '/settings/about',
     builder: dummyBuilder,
     redirections: [
@@ -56,11 +67,11 @@ class TestDestinations {
   static final catalogTransit = Destination.transit(
     path: '/catalog',
     navigator: TestNavigators.catalog,
-    builder: (context, parameters, child) {
+    builder: (context, parameters, childBuilder) {
       return Column(
         children: [
           const Text('Catalog'),
-          Expanded(child: child),
+          Expanded(child: childBuilder(context)),
         ],
       );
     },
@@ -101,11 +112,8 @@ class TestDestinations {
 }
 
 class TestNavigators {
-  static final catalog = NavigationController(
-    destinations: [
-      TestDestinations.categories,
-      TestDestinations.categoriesBrands,
-    ],
-    tag: 'Catalog'
-  );
+  static final catalog = NavigationController(destinations: [
+    TestDestinations.categories,
+    TestDestinations.categoriesBrands,
+  ], tag: 'Catalog');
 }